Fix case-sensitive opml feed parser
The OPML spec seems to be ambiguous on the subject of case-sensitivity.
As a result, OPML feeds that are exported from podcatchers may have
uppercase attributes (i.e., `URL` vs `url`). This patch will find
attributes of `xml.etree.ElementTree.Element` objects in a
case-insensitive way which side-steps the problem.